+++ src/message.rs
@@ -1,11 +1,12 @@
 use crate::actor::{ActorTrait, Handler};
 use crate::actor_ref::ActorHandler;
 use serde::Serialize;
+use std::any::{TypeId, Any};
 
 pub trait MessageTrait: Send + Sync {}
 
 pub trait MessageEnvelopeTrait<A>: Send + Sync {
-    fn handle(&mut self, actor: &mut A) {}
+    fn handle(&mut self, actor: &mut A) -> MessageType;
 }
 
 pub struct MessageEnvelope<A>(Box<dyn MessageEnvelopeTrait<A> + Send + Sync>);
@@ -21,7 +22,7 @@ impl<A> MessageEnvelope<A> {
 }
 
 impl<A> MessageEnvelopeTrait<A> for MessageEnvelope<A> {
-    fn handle(&mut self, act: &mut A) {
+    fn handle(&mut self, act: &mut A) -> MessageType {
         self.0.handle(act)
     }
 }
@@ -38,9 +39,25 @@ where
     M: MessageTrait + Send + 'static,
     A: Handler<M>,
 {
-    fn handle(&mut self, act: &mut A) {
+    fn handle(&mut self, act: &mut A) -> MessageType {
         if let Some(msg) = self.msg.take() {
+            let msg_type_id = msg.type_id();
             act.handle(msg);
+            if msg_type_id == TypeId::of::<StopMessage>() {
+                return MessageType::StopMessage
+            }
         }
+        MessageType::Unknown
     }
+}
+
+#[derive(Clone)]
+pub struct StopMessage {}
+
+impl MessageTrait for StopMessage {}
+
+#[derive(PartialEq, Clone, Copy, Debug)]
+pub enum MessageType {
+    Unknown,
+    StopMessage
 }